phase-IV/backend/requirements.txt
fastapi==0.110.3
sqlmodel==0.0.16
pydantic>=2.12.3
pydantic-settings>=2.5.2
uvicorn>=0.31.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pytest==7.4.3
httpx>=0.27.1
python-dotenv==1.0.0
pyjwt>=2.10.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
# AI & MCP dependencies
openai-agents==0.8.1
mcp==1.26.0
tiktoken==0.12.0